Squash Chromium prebuilt
 vendor/omni: Add chromium_prebuilt.sh

This script will pull the necessary libs from a complete build and generate proper
makefiles in order to support using a prebuilt Chromium. Files are placed in
prebuilts/chromium/-devicenamehere-, and are not tracked by git, so they are completely
out of the way.

PS3:
-replace git -C with params compatible to old git versions

Update chromium_prebuilt.sh for 5.0

 Fix-ups for prebuilt Chromium

- Make copying far more reliable, even on non-clean builds
- Use $OUT in script in all cases, apparently get_build_var OUT_DIR isn't reliable.
envsetup.sh, please -_-

Change-Id: I4b0f11a8f0c4c4762fd1a8d0f6cdc54abadaf15d
Signed-off-by: k2wl <kar2wlan@gmail.com>